backend/requirements.txt
fastapi==0.108.0
uvicorn[standard]==0.25.0
paho-mqtt==1.6.1
sqlalchemy==2.0.23
python-dotenv==1.0.0
websockets==12.0
